requirements.txt
# Core dependencies
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Email processing
mailbox>=0.4
python-dateutil>=2.8.2
# Removed email-parser as it's not compatible with Python 3.12
# We'll use standard email and mailbox libraries instead

# NLP and embeddings
transformers>=4.36.0
sentence-transformers>=2.2.2
spacy>=3.7.0
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
nltk>=3.8

# Vector storage and search
elasticsearch>=8.10.0
elasticsearch-dsl>=8.0.0

# RAG and retrieval
faiss-cpu>=1.7.0  # or faiss-gpu if you have a GPU
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2

# Visualization
plotly>=5.18.0
networkx>=3.1
pyvis>=0.3.2
streamlit-modal>=0.1.0
streamlit-aggrid>=0.3.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0